Add data tests (#52)
Add all files in
https://github.com/IrreducibleOSS/zCray-lang/tree/main/data (except
parity), and add tests, opcodes and fixes to make them work, if we omit
2-bit complement (not yet implemented).

This PR includes a change in the `schoolbook_multiplication` method so
that it can be used for MUL (u32 * u32) as well as MULI (u32 * u16).

+++ assembly/src/parser/instruction_args.rs
@@ -71,7 +71,9 @@ impl std::str::FromStr for Immediate {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let is_field = s.ends_with('G');
         let s = s.trim_start_matches('#').trim_end_matches("G");
-        let int_val = i16::from_str(s).map_err(|_| BadArgumentError::Immediate(s.to_string()))?;
+
+        let int_val =
+            i64::from_str(s).map_err(|_| BadArgumentError::Immediate(s.to_string()))? as i32;
         if is_field {
             let v = BinaryField32b::MULTIPLICATIVE_GENERATOR.pow(int_val.unsigned_abs() as u64);
             if int_val < 0 {
